Update HTML/CSS/JS for layout and responsive navbar
This is long overdue. The CSS and JS added is for a responsive navbar
when the screen is at smaller sizes. Using aria-expanded and
a new data-visible attribute helped create the code needed in both
JS and CSS for an onclick event to reveal and hide the navbar.

A hamburger menu svg and close panel svg were placed using CSS
to give the user an area to click to make the navbar hide or
reveal itself.