Finish HTML and CSS templates
Added the layout and styling for the portfolio section and contact section

The styling for the intro section's title was changed to have a
animated gradient clipped to the transparent text.

A background was placed in the side-navbar to give it some more
interest.

The mobile layout was fixed. The grid-template-areas was changed
in the section intro grid so that it stacked up nicely.

The content was all centered in mobile layout using justify content
on the grid parent element, section-intro.